Add Ceph migration procedure
This commit introduces the Ceph adoption document that will be used to
document the multiple scenario involved.
Scenario 1, the very basic one, has been added, and it proves that mon
and mgr can be migrated keeping the same IP address. By doing this we
should be able to preserve the running workload and "externalize" a
TripleO deployed Ceph cluster.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>