Move secrets and vars out of inventory
Based on a discussion i'm moving the secrets into their own file, and
i'll do the same for variables. This loses the advantage of having a
single file with all environment info, but may be more suitable for
CI, e.g. writing out the vars/secrets files via automation without
having to watch out for YAML indentation in the inventory.