Document Development Environment based on install_yamls
This version of the documents describe how to use
install_yamls/devsetup to configure edpm-compute-0
with isolated networks which can access the NG control
plane on k8s. It then covers installing OpenStack Wallaby
using TripleO Standalone on the edpm-compute-0 node with
network isolation and Ceph and provides example environment
files. It then explains how to add an NG deployment to
the environment and leaves the reader linked to the
adoption procedures they can follow in the new environment.

Signed-off-by: John Fulton <fulton@redhat.com>