ovn: remove troubleshooting note about ovn db version mismatch
The original note was added due to OSP 17.x deploying OVN version that
was larger than what was included in new containers. Now that OVN is
bumped in the containers, this situation can't ever happen.

Also, we just added a step to OVN page where we upgrade the old database
schema version to the latest from the new ovsdb-server, so the mismatch
is impossible in the other direction (old_version < new_version).

Hence removing the note as not applicable to any supported environments.