Dev Env: Include wget commands
While following the Development Environment process it's a bit
unconfortable when we suddently have to download a file for the next
command, and we may even miss that we have to download it if we are
going fast just running the terminal commands.

This patch just adds the `wget` commands to download the files in the
right locations to facilitate the process.