Dev Env: Route mariadb VLAN
The adoption process requires that the DB is accessible from the host
where we run the commands, but this is not done in the development
environment process.

This patch adds the commands necessary to route the traffic and be able
to run the `mysql` commands as the adoption procedure indicates.