Planning the deployment
This patch adds an initial step for the adoption, which is the planning
phase, just like we have in the normal OSP documentation.

This planning document is not complete, but at least serves as an
introduction and has a couple of TODOs for others more knowledgeable is
those parts to complete it.

It also adds a specific document on how to select nodes to deploy
OpenStack services, which is important for services such as cinder where
specific hardware (like FC HBAs) may not be available in all the
OpenShift nodes.