KeyStone: Use OpenShift deployed openstack client
We have never asked to install the openstack client yet we somehow ask
to use it, in the keystone adoption, without explaining how to install
it.

This patch changes how we run the `openstack` client to use the one
deployed inside OpenShift, which also serves as a way to illustrate how
we can easily run openstack commands on the new deployment.