Glance: Fix secret and show alternative creation
This patch fixes the Secret example in the Glance adoption that has the
wrong name.

The secret name used is "ceph-client-con" but then the Glance deployment
uses "ceph-conf-files".

The patch also adds an alternative mechanism that doesn't require us to
manually craft the Secret and instead create it from the files
themselves.