Glance: Fix the file with the config
The current Glance adoption process is using the old 01-custom.conf
filename to look for the new service configuration, but now the right
name is 02-global.conf, this patches fixes this issue.

The patch also provides full commands to check the contents of the
container using oc commands instead of assuming that the human operator
following the procedure will figure out they have to list the pods and
then go into one of the glance ones to run those commands.