Refactor the test suite according to docs improvement PR
This updates the test suite to follow what was edited in docs in PR
55:

https://github.com/openstack-k8s-operators/data-plane-adoption/pull/55

Mainly this splits OpenStack service stopping into its own role (as it
has its own doc now) and renames 'EXTERNAL' env vars to 'SOURCE'. PR
55 did this rename for MariaDB only, this commit also does the same
edit (in both docs and tests) in the OVN adoption procedure to keep
consistency in variable naming.