extend mariadb_copy to split out nova cell1 into a seperate db
This change modifes the default contolplane CR to deploy two mariadb
instances, the common openstack db and a seperate openstack-cell1
db. The later will be used to adopt the cell1 db called "nova" into
a seperate db instance. this logic is generic and has replaced the
prior rename logic for neutron.

This change extends the extends the mariadb_copy doc and test
to split up the resotre of the db into nova and non_nova dbs.
the non nova dbs are all adopted to the shared openstack db instance.
the nova_api and nova_cell0 db are adopted to the openstack db
and the nova db is adopted into the openstack-cell1 db.