/bin/sh is not /bin/bash
The ansible shell and command modules use /bin/sh not
the login shell of the remote user.
command is always /bin/sh where as the shell module
can sepcify a executable to use as the shell.

This change correct some of our missuse of the shell module
by explictly using /bin/bash. This is done by setting the
module_defaults for the playbooks to /bin/bash

The adoption repo is explcilty trying to test the procedure
for adoption so we avoid using ansible modules and mirror
the actions that users will perform as closely as possible.
Since we intened to be persciptive and have the users follow
the procedure from a bash shell we mimic that by defaulting
the executable to /bin/bash

This change fixes a race between wait for the keystone
pod to be ready and trying to use the openstackclient
pod.

This change also address command not found issues with
the use of a bash alias when the test are run on a host
that does not use /bin/bash as /bin/sh

When using an alias we already used ${BASH_ALIASES[openstack]}
to avoid command not found if invoking the alias via xargs
this change extends that to all useses of aliases.
While it is not clear why this appears to be requried
it may be related to the bash version or other envionmental
factors. As ${BASH_ALIASES[openstack]} seams to be more robust
all implict use of aliases are not made explict in the test code.
The docs have been updated to reflect that you should transform
implict aliases to explicit aliases in the test roles.