Add sample env file to persist adoption settings
Provide a sample env file to recover settings after rebooting the
hypervisor. It is mostly useful for resuming of the interrupted manual
steps, when following the documentation. Especially, when the
hypervisor is a VM in a cloud provider, like PSI.

The helper env file ensures that the already deployed crc and edpm
standalone host (compute 0) are running, logs in to crc, exports env
vars and makes aliases for openstack and podman CLI to exec it on the
standalone edpm host via ssh.

This saves time, but requires the documented manual steps and the
sample env file contents to be in sync. There is already such a
requirement for the documentated scripts and test suits. We could
leverage code generation in the future to update docs and tests
automatically, when the scripted commands get updated in a commit.
While the scripts and env files, like that sample env file, should
then become the only source of truth.

NOTE: Stability of CRC and EDPM hosts after restarting in unrelated,
but it shows reliable enough for simple adoption scenarios.

Additionally, follow-up multi-cell DB adoption patch with service
IP selection and trivial typos fixing.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>