Remove remaining container images from test roles
The two containerImage parameters removed here are the only ones
remaining in the test roles. For normal usage, this is defaulted in
the operators and we shouldn't need to set it.

For periodic CI, these values are overriden with particular non-latest
container URLs via the kustomization mechanism.

So i think these shouldn't be necessary and can be removed too.

Related:

https://github.com/openstack-k8s-operators/data-plane-adoption/issues/110
https://github.com/openstack-k8s-operators/data-plane-adoption/pull/165