Adapt the data-plane-adoption doc/tests to the new GlanceAPI CRD
We recently introduced the ability to deploy a list of glanceAPI
in the OpenStackControlPlane. This feature is not required in the
adoption context, but it's critical to adapt the Glance deployment
to the new api.
It might be relevant in the future as long as DCN environments will
be adopted.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>