[edpm]Temporary fix for stable compute UUID
The temporary compute UUID feature exits in Antelope but the backport of
it[1] to OSP 17.x hasn't been landed yet. When we start pre-generating
the compute UUID during greenfield deployment as part of
https://github.com/openstack-k8s-operators/edpm-ansible/pull/498 the
adoption will break if there is no compute UUID file exists on each
compute node. When [1] is released in 17.x the compute UUID file will be
created by the nova-compute on the 17.x side. Until that we create the
file right at the start of the EDPM adoption.

[1] https://code.engineering.redhat.com/gerrit/q/topic:stable-compute-uuid

Implements: OSPRH-117