Make local-storage SC options more flexible
Make storage_class_name variable controling which
source SC should be used to create a local-storage SC from it.

When it points a non-existing SC, use the default CRC SC as
the local-storage SC source. Otherwise, create the local-storage
SC frin that source SC.

Always refer that local-storage SC by name in sample/oscp
CRs, and intermittent data migration pods. That is, because
changing the name of SC in CRs and pods specs is complicated and
error-prone, and requires adjustments in multiple places, including
oscp spec.glanceAPIs.

Remove dependencies on install_yaml make crc_storage* targets
and switch storageClasse in the adoption related specs to use
the default CRC sc. Later on, this default may become replaced
with LVMS, or whatnot.

Replace the make targets for storage to direct ocp commands
and PVC targets for it.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>