Rebuild docs if any of the sources change
Previously the `html` make target for docs only had `main.adoc` as the
source. That means if we were including some assemblies from
`assemblies` directory, running `make html` would not rebuild the
docs. We're now trying to track all sources in the `ALL_SOURCES`
variable and rebuild the docs if any of the sources change.