Add watch-html targets for docs
These targets will set up inotify watches for docs sources and if any
of the source files changes, they will re-render the docs using the
`make html` target.

There is a 0.5 second sleep between the inotify trigger and the docs
rendering. This is to ensure that if multiple files were edited and
"Save All" function of a text editor was used, there is a bit of time
to allow all of the files to finish saving before we start rendering
the docs.