Tests: Cinder volume and backup support
This patch adds the cinder volume and backup services to the tests.

For now only Ceph is supported in both services, and the backup can also
be deployed with swift.

By default none of these cinder services are deployed and we must use
`cinder_volume_backend` and `cinder_backup_backend` to deploy them, as
illustrated in `tests/playbooks/test_with_ceph.yaml`.

For cinder volume and backup services we'll not only confirm that the
pods are running, but also that the services are reporting as being
`up`.

Jira: OSPRH-2442