Check pull requests that docs rendering works
Until now we ran docs build & push only after a push to the main
branch. Newly we'll run the docs build workflow both on a push to the
main branch and on pull requests. We'll only run the last task (acutal
push to Github Pages) on a push to the main branch, not on pull
requests.