Tests: Fix Cinder test with no volume or backup
The logic to detect when cinder volume and cinder backups are running in
the "wait for Cinder pods to start up" task of the cinder_adoption role
fails when there is no volume or backend, because the logic is faulty.

This patch changes the logic so that either the variable that specifies
the backup and volume backend is empty or the `oc` command to check for
the status of the pod succeeds.