Move 'Configuring a Ceph backend' right after initial control plane creation
Previously the Ceph backend configuration was after Keystone and
Neutron adoption, before Glance adoption. While it makes some sense to
have it before adopting the first service that makes use of the Ceph
backend, inserting it in the middle of service adoption docs breaks
the adoption flow too. The best place to configure the Ceph backend is
probably right after the intial control plane deployment. This means
we'll have an uninterrupted sequence of "<$service> adoption"
steps later in the workflow.