Adoption of the ovn-controller on the EDPM nodes
This patch adds docs updates and test changes to cover adoption of
adoption of the ovn-controller running on the EDPM nodes.
There is no really many changes required to make it working. The most
important thing is to make sure that bridge mappings configured on the
node are the same before and after adoption.

Regarding tests, it only makes sure that test VM created before adoption
is reachable before and after adoption.

Closes: issue/OSPRH-5015