Adoption of the neutron-ovn-metadata-agent on EDPM nodes
This patch adds short note about adoption of the
neutron-ovn-metadata-agent in the adoption user's doc.
Adoption of this agent don't really require any aditional action during
the adoption process.

This patch also adds simple check in the test framework to make sure
that agent is alive after the adoption is finished.

Closes: issue/OSPRH-2438