Patch glanceAPI using per-backend files
Instead of passing an inline string to patch the controlplane, rely on
per-backend files (because the config is pretty much static) that can
be called from the role. This allows to improve readability and it helps
in terms of maintenance, and it also avoid conflicts in case we use
golang templating within the customServiceConfig option.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>