Re-add cinder-cfg.py script
On commit fba8ae0129655795d5481a1434265a2bc29d8342 the cinder-cfg.py
script was removed when cleaning up old Markdown/MkDocs files, but this
file is still being referenced in the documentation, so we'll get a 404
error when trying to follow the link in the rendered docs.

This patch adds the file back as it was right before being removed.