Neutron SR-IOV agent adoption
This patch adds adoption of the neutron-sriov-nic-agent on the edpm
nodes.
As we run our tests on the virtualized environment, there is no really
way to spawn VM which uses SR-IOV port and test if this VM works after
adoption. This will need to be tested in some other layer of our CI.
This patch adds documentation about adoption of the SR-IOV agent and
simple test which will make sure that neutron-sriov-nic-agent is ALIVE
after adoption.

Depends-On: https://github.com/openstack-k8s-operators/install_yamls/pull/755

Closes: OSPRH-2433