Translate the ironic-inspector db name
In puppet-ironic the database name for ironic inspector service
used a dash (`-`). In the mariadb-operator we cannot use a dash
in the database name, it will cause a SQL syntax error when creating
the database.

Add ironic-inspector to the db_name_map.

This aligns with the expected database name in ironic-operator.

Also store the SQL command in a variable, and echo that to make the SQL
command available in logs/output.