Show upstream version in github preview
The github doc preview is missing sections of the doc as the build
variable is not defined in that env and therefore all our conditional
upstream/downstream ifeval blocks are hidden.

This patch changes the upstream conditions from:
    ifeval::["{build}" == "upstream"]
to
    ifeval::["{build}" != "downstream"]
so the upsteam blocks are visible even if the build variable is not
defined.