Document details about network adoption when using a new subnet for CP
Using a new subnet for CP requires that we set up link local routes on
both OCP workers and tripleo nodes. For OCP, it's done in
`NetworkNodeConfigurationPolicy` CRs. For tripleo nodes, it's done via
`net_config.yaml` files, listing additional `routes` with `next_hop` set
to {{ ctlplane_ip }}.

Once this is done, the next step is to set EDP NodeSet to retain the
routes in net_config; as well as configure ssh range to allow access
from ansibleee pods.

This procedure was validated using
https://github.com/openstack-k8s-operators/install_yamls/pull/774

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>