Add Swift data migration docs and test
Describes and tests migrating Swift data from existing nodes to new
nodes using PVCs as backends.

Due to the length of the section and the possibility to migrate data
without downtime after all other services, this has been put into its
own assembly.

Tests are put into a new role and not enabled by default, as these take
quite some time to finish. It also requires a standalone deployment with
at least two replicas for Swift. Standalone with a single Swift replica
does not enable replication processes, but these are required for the
migration.

If using install_yamls this can be enabled with the following
environment variable:

  SWIFT_REPLICATED=true make standalone

And tests can then be run after adoption finished:

  make test-minimal
  make test-swift-migration

Tests are using "forced_rebalance" instead of "rebalance"; this is to
avoid a minimum waiting time of one hour between rebalances in the
tests.