Add CephFS-via-NFS migration guide
This setup concerns an OpenStack-adjacent data plane
service called "ceph-nfs". On RHOSP 17/Wallaby,
this service encapsulates an NFS-Ganesha server
and a VIP and is orchestrated by pacemaker and systemd.
When adopting to RHOSP 18, we no longer support deploying
this service on RHOSP. We document how deployers can
replace this service with a new clustered NFS service
natively supported on Ceph.

Implements: OSPRH-2238