do not create a seperate nova service for post ffu
the previous workflow created a new configmap and
dataplane service to deploy nova in a second deployment
post ffu. this was inherintly buggy as the node set still
refernced the orginal dataplane service and the old configmap.

That means if you did a deployment in the futrue it would revert
to the ffu config.

This change correct that by updating the workflow to modify the existing
configmap and redeploy the orginal service. This is more inline with
how adoption is intended to work as we dont intend multiple datapalne
services for the same openstack service to be used against the
same nodeset.