ironic standalone - disable neutron-sriov-nic-agent
The standalone deployment fails when EDPM_COMPUTE_SRIOV_ENABLED
is `true` in combination with `STANDALONE_COMPUTE_DRIVER`.

Update the developer docs with instructions to disable it.

!! This might be a bug in tripleo?
We should possibly set `nova::compute::pci::passthrough` to an
empty list in [1] or possibly return an empty list at the end of
the user_passthrough_config() in [2].

[1] https://opendev.org/openstack/tripleo-heat-templates/src/branch/stable/wallaby/deployment/nova/nova-ironic-container-puppet.yaml
[2] https://opendev.org/openstack/tripleo-heat-templates/src/branch/stable/wallaby/deployment/neutron/derive_pci_passthrough_whitelist.py#L241