[CI]Wait for operator restart after csv patch
The ansible runner image is built by the CI job and the
ansibleee-operator csv is patched with the ENV variable to use the
freshly built image. This triggers an operator restart which can take
more than couple of seconds (we saw 30 secs in some cases). So the CI
job needs to wait for the operator to restart before moving forward
otherwise it is possible the resources created later will be reconciled
by the old operator and therefore uses the old runner image.