Remove ovn_match_northd_version from samples
`false` is the default value; and we are not relying on this ovn
"feature" when upgrading ovs. (it is needed only when upgrading using
so-called `fail-safe upgrade` procedure [1], which means upgrading
skipping an LTS release. This is not what we do anyway.)

Nor it makes sense to migrate the value from adopted hosts - it should
always be `false`.

[1] https://docs.ovn.org/en/stable/intro/install/ovn-upgrades.html#fail-safe-upgrade