Add variable to skip patching ansiblee csv
In some ocasions, like beta testing we want to skip patching the
ansiblee csv to set latest, and instead rely on the image that the
operator has by default. This change introduces a variable that we can
set to skip the related tasks in dataplane_adoption role. By default the
variable is set to false so we keep the current behaviour.