Add wait_daemons tasks before moving to the next phase
This patch adds a better orchestration of the time that we need to wait
before moving to the next phase. In particular it sends a query to the
ceph cluster and gather the daemons (per type) that are going to be
deployed. It waits for the status to be "running" before moving to the
next phase.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>