Reorganize Ceph assemblies
Ceph assemblies are now better reorganized to follow a simple rule/struct.
A main ceph-cluster migration assembly is included in main, and it contains
a quick intro and the (ordered) list of procedures (including the cardinality
section that is critical here and will be improved in a follow up patch).
This way the ceph doc is very easy to access and maintain. There are also
fixes to wrong references (e.g. horizon != Ceph dashboard).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>