docs: Add section about OVN gateway node adoption
The procedure is a truncated version of Compute node section.

The main difference is in the list of services deployed that doesn't
include libvirt and nova-compute-*.

Also, edpm_enable_chassis_gw is set to true to serve as OVN gateway
node.

Another difference is that sriov service is never deployed on Networker
nodes. Metadata service is or is not deployed, depending on the presence
of SR-IOV ports on peer Compute nodes.

The patch also updates OVN data migration section for more clarity, and
to mention that hosting OVN gateways on OCP nodes is not as robust as on
EDP nodes.

Test suite is updated to include a new variable  - compute_adoption - to
allow to disable compute specific adoption steps (set to `true` by
default.) The configuration was validated with a standalone tripleo
setup deployed by install_yamls (`make standalone`).