Add swift into the deployed OpenStackControlPlane for scale-in check
As described in the related bug, we are not specifying swift in our
initial controlplane deployment. The default for swiftStorage is to
have replicas: 1 so when we patch the controlplane with replicas: 0
it triggers the 'no valid scale-in' validation.

Related-Bug: https://issues.redhat.com/browse/OSPCIX-320
(cherry picked from commit ca3fcb2ec050e99a1e807dd18b7f6357f0db7d31)