Move manila cephfs spec definition to a template
Instead of hardcoding the manila_cephfs config to the ansible task,
this change organizes it into a template, that is rendered and applied
during the adoption.
The template adds an initial support for NFS backend, in preparation to
a follow up patch where the NFS adoption test will be added.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>