Update custom container images
The way custom images are specified has changed, and now it no longer
uses the `containerImage` file in the cinder sections but instead needs
to be specified in the `OpenStackVersion`.

This patch updates the cinder and manila docs to the latest code, as
well as the cinder-cfg.py helper code.