Test: Wait for cinder pods to be `Ready`
The current code waits for cinder pods to be `Running`, but that doesn't
mean that the pod is ready, as it may be running just one of the
multiple containers.

This patch changes how the cinder role waits for cinder pod services,
and now waits for them to be `Ready` instead.

Jira: OSPRH-8814