Add mon_map as a variable override and make the mon code more modular
Sometimes it's easy to define, as overrides, immutable data structures.
Currently, the mon_ipaddr is gathered from the ceph cluster, but the
mon_dump might change from run to run, invalidating the use case we'd
like to solve.
This patch adds the "mon_map" struct as an override that can be provided
by the administrator to statically define the src hosts information
related to the mon migration.
In addition, the backup set of tasks are moved into a new file: it's
easy to read/manage and removes duplicated code.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>