Refresh mon config when ceph_keep_mon_ipaddr is not true
If mons ip addresses are not preserved, the tmp client needs fresh
information about the current mons, otherwise it won't be able to
connect to Ceph anymore at the third iteration (after controller-0
and controller-1 are gone).
This patch ensures we can get fresh information for the client after
each iteration.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>