Add a note regarding neutron-dhcp and baremetal
This patch is adding a note in the section that explains how to enable
neutron-dhcp about using it for baremetal provisioning.

For baremetal provisoning, either the built-in DHCP from OVN or
Neutron DHCP Agent can be used to provide PXE and DHCP for the
baremetal nodes. By default, the OVN built-in DHCP is used.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>