Extend mon label to CephStorage nodes during mon migration
When mons are migrated we can't risk to lose quorum during the
replacement procedure. For this reason this patch adds a step in the mon
migration procedure to extend the mon label to the rest of CephStorage
nodes (target nodes hosting OSDs). By doing this we can make sure
temporary mons are deployed, preserving the quorum during the entire
process.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>