Verify if the existing test VM instance is running
Add an extra verification of the test VM instance created
on the source cloud, before adoption started.

This makes that failure mode verification more explicit,
and the test suit to fail earlier.

Without that check, the next check to verify if the test VM
instance can be stopped will fail, eventually. That is OK
as well, yet obfuscates the real cause of an important failure
mode - a test instance must survive adoption without downtime.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>