Change ready condition for aodh
The pod lifecycle doc mentions about pod phase[1] and pod conditions[2].
Acoording to the phase states, a pod gets into "Running" state even when
one of the containers is running or in the process of starting.
This is not accurate since we desire the pod to be running all the containers
before proceeding with the deployment.
As per pod conditions, a Ready state is when a pod is ready to serve request
(all containers running) and can be added to load balancing pool of services.

Hence, the pod condition "Ready" is a better check for pod readiness rather
than checking the "Running" phase of a pod.

Same changes have already been implemented in the cinder_adoption role by
PR[3].

[1] https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
[2] https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-conditions
[3] https://github.com/openstack-k8s-operators/data-plane-adoption/pull/547