Add Manila CephNFS support
This patch introduces a set of basic tasks to perform the adoption of
Manila with a CephNFS backend.
Before performing the actual adoption, a new CephNFS cephadm based
cluster must be created. For this reason, the manila_nfs tasks require a
few parameters as input:

1. A new CephNFS VIP where the CephIngress daemon (made by Haproxy and
   keepalived) is created;
2. A set of target nodes where the "nfs" label is added: these nodes are
   supposed to host the new CephNFS cluster and must be different from
   controller nodes that are going to be decommisioned;
3. The TripleO managed Ganesha VIP: this input can be actually retrieved
   by the existing TripleO deployment, but we can gather this input as
   part of the next iteration on this patch

After the new CephNFS cluster is created, it is possible to build
manila-share with the proper configuration.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>