Add TripleO requirements playbook
This playbook represents a requirement for some storage services (e.g.
Manila, but it can be extended to Glance NFS in the future) to prepare
a TripleO based environment to perform the adoption. Manila, when a
ceph-nfs backend is used, requires a cephadm based NFS cluster, and the
goal of these tasks are:
1. Propagate the StorageNFS network to the Ceph target nodes
2. Deploy a CephNFS cluster to the specified VIP

This is in line with the documentation and by executing this preliminary
playbook we're able to simplify the Manila adoption process.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>