Validate Cinder operations
After the adoption, we check for the correct status and size of the
Cinder resources but don't validate that operations still work on them.
This patch adds the following operations:
1. Detach 'disk' volume from 'test' server
2. Revert to snapshot 'snapshot' to 'disk' volume
3. Attach 'disk' volume to 'test' server

This will help verify that the resources are in consistent state before
and after the adoption.

NOTE: We are skipping restore backup since it's	failing	even with 20
retries	with 10 second delay. Will add that in a follow up PR after
resolving the issues.