Don't use/verify Cinder capabilities if no backend
Cinder volume and backup should have supported storage backends
configured in order to use and verify volumes/snapshots with a
test VM that we start before adoption.

When the volume backend is not supported, do not attach volumes
and do not create snapshots for the test/bfv VMs.

When the backup backend is not supported, do not create volumes
backups.

Neither verify, if those are still present post-adoption.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>