Add Ceph post configuration section
This patch adds the Ceph post migration procedure.
At the moment the only thing that should be reconfigured is the Ceph
Dashboard failover, but this guide might be extended according to the
feedback we receive from the field.
The post-rbd verification is also improved to verify mons are running
using the old IP addresses.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>