[ovn] Add iptables rules on controllers for scenario A
If the adoption is done with different subnets between the
wallaby environment and the podified controlplane, it will fail
on the ovn dump database since on the wallaby environment there's
a firewall rule which only enables traffic to the NB/SB DB if it
belongs to the same wallaby internal_api network.

If the adoption has different subnets (wallaby internal_api !=
podified internal_api) a new firewall rule is needed to accept all
traffic comming from podified internal_api.

Resolves: OSPRH-10895