Do not fail when adoption is performed using Ceph Reef
Quincy was used to develop the migration procedure. However, as per the
official documentation, Reef is the required version. This patch allows
to perform the Ceph migration with Reef, and it adds a boolean to skip
this validation in case we need to test a version != from Reef (mostly
used for previous versions).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>