Cinder: Set default volume type
In the standalone 17.1 deployment, we set the default volume
type for the deployment.
The same type is not set in RHOSO deployment hence we see
failures in some jobs[1].
To mitigate this issue, we fetch the ``default_volume_type``
field from 17.1 deployment's cinder.conf and patch it in
the openstackcontrolplane CRs cinder section.

[1] https://logserver.rdoproject.org/28/55128/1/check/periodic-adoption-multinode-to-crc-no-ceph/f3e425b/controller/ci-framework-data/tests/test_operator/tempest-tests/stestr_results.html