Use a more generic rc check in os-diff tasks
Instead of checking for a rc of 1, check if it's not 0. For some reason,
while testing unigamma, the task to check for go fails with a rc of 2,
which fails the tests because it skips the task to install go, even though
it's not installed.

(cherry picked from commit 7cf0408e6cad5a5f44dcd03deb9cbbacb275dc5f)