Basic ping test for test_with_ceph.
When both `prelaunch_test_instance` and `ping_test` are true, starts a
ping during `development_environment` and stop it in
`_stop_ping_test.yaml`.  Depending on the threashold defined, it will
ensure that we were able to ping the instance' FIP during the entire
adoption process.

Currently only activated for jobs that leverage `test_with_ceph.yaml`.

When we got some ping failure but still are under the threshold we add
a warning to the output, without failing. That might be useful to
debug the problem later.

We also convert the ping log with human readable date and time for
easier debugging when the ping test fails or is in warning.

Test is deactivated by default and ping loss should be 0.  All those
can be configured in the job.

(cherry picked from commit 7b19bd00251c209fa3f243ecb5410ae29933c04b)