Block Storage: Correct deployment steps
To deploy block storage service i.e. cinder, we should first remove
the service entries from old deployments to avoid any issues in
the new RHOSO deployment.

For that we need to rewrite the steps to:

1. Deploy cinder-api pod
2. List all scheduler and backup services from old deployment
3. Remove the old scheduler and backup services
4. Deploy scheduler, volume and backup services
5. Perform DB migration

This will ensure that our RHOSO deployment is fresh and does not
contain traces of errors from services of old deployment.