Add noop job to jobs-layout
In order to get proper feedback from the rdoproject.org/github-check,
the noop job is added. This job should be kicked even if other jobs
are not started as irrelevant-files are detected in change.
Without noop job rdoproject.org/github-check is hanging with waiting for
status to be reported message.