Fix post processing process for ping test.
We have certain post processes that parse all the file that ends in
`.log` in the adoption `log` directory.  It expects an Ansible log
format.

So in order to not confuse that process, the ping files are renamed to
end with `.txt`.

Furthermore, for some reason, the files were not readable by that
process. Even if now they will be excluded from the processing, it's
best if they are world readable.

(cherry picked from commit bf1541b2bf224de2c53016521bc86d2150a71498)