Fix: Cleanup old cinder services
Somehow the services are not being captured with the volume service
list while grepping for the "down" services.
One possibility is that the services were in "up" state when we
migrated the DB from 17.1 -> 18 and it takes time to update the
DB to report them as "down".
The straightforward way would be to cleanup all old cinder-scheduler
and cinder-backup service entries since we haven't deployed the
RHOSO setup yet.

+------------------+------------------------+------+---------+-------+----------------------------+
| Binary           | Host                   | Zone | Status  | State | Updated At                 |
+------------------+------------------------+------+---------+-------+----------------------------+
| cinder-scheduler | standalone.localdomain | nova | enabled | down  | 2024-11-13T14:21:59.000000 |
| cinder-backup    | standalone.localdomain | nova | enabled | down  | 2024-11-13T14:21:59.000000 |
| cinder-volume    | hostgroup@tripleo_ceph | nova | enabled | up    | 2024-11-13T15:34:40.000000 |
| cinder-scheduler | cinder-scheduler-0     | nova | enabled | up    | 2024-11-13T15:34:45.000000 |
| cinder-backup    | cinder-backup-0        | nova | enabled | up    | 2024-11-13T15:34:48.000000 |
+------------------+------------------------+------+---------+-------+----------------------------+