Switch precedence of secrets.yaml before vars.yaml
In automated tests, CI jobs may need to redefine vars like
ipa_ssh, or controller1_ssh, which are stored in secrets.yaml
and always taking the most precedence. To w/a that, rdo-jobs
creates inline hacks for secrets.yaml, which still does not
work when we want to override it via adoption_extra_vars.

Switch the precedence of included secrets then vars, to allow
overriding anything in adoption_extra_vars.

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>