Use MariaDB/Rabbit verification commands that will make issues visible
The previous version with '| grep Running' may be useful in tests to
force a non-zero return code and abort a block of code, but in docs
the grepping will hide anything that is not "Running", so we actually
do not see if there are any issues.

Therefore in the docs the 'grep' is now removed and the admin is told
to check if all status commands print "Running".