[18.0-fr2] Fixed incorrect grep in octavia adoption
Use "\|" instead of "|" in grep, previous regexp didn't work as expected
because grep considers "octbr|enp6s0.24" as a full-string.

It fixes a bug if the octavia bridge and interface are already present
in the NNCP, this NNCP might have been patched again with a different
configuration.

Also skip the update of nncp-dns-* NNCPs which are not related to
the octavia network config.

JIRA: OSPCIX-747
(cherry picked from commit d391d6b7140fa12152cc7a67bf9bc86f62647bab)