Fix ping test failures in pre_launch.bash
The `ping -c4` is sometimes not enough time for the instance to start
responding, and the ping test fails. We shouldn't just increase the
number for `-c` because that would prolong the test even if the ping
is already responding. We now use a small bash loop to quit at the
first ping success or error out due to (now considerably longer)
timeout.

Implementation note: ((TRIES++)) returns 1 when TRIES=1, so we work
around it by `|| true` to not crash our script which uses `set -e`.