Refactor documentation for enabling TLS-e during controlplane creation
Add tlse spec with disabled tlse to example controlplane CR and
document in note how tlse can be enabled if source OSP has TLS-e enabled.
This is in sync with test suite and EDPM adoption part, where we have
tlse disabled and document what to do to enable it.

In review Katie asked 2 notes to be moved below the code block.

Resolves: https://issues.redhat.com/browse/OSPRH-14786